chore: move renderListingsByCategory to appropriate folder
